Trim fence results
By trimming we avoid any special carriage returns showing up in the
input. Also when a command adds newlines the cleanup call was breaking
since for now it deletes by a paragrah. If this ends up causing issues I
am going to have to find a better way to model the results output area.